+++ src/api/objects/toot.ts
@@ -631,7 +631,16 @@ export default class Toot implements TootObj {
 
     // Returns true if the toot should be re-completed
     private shouldComplete(): boolean {
-        return !this.completedAt || (ageInMinutes(this.completedAt) > Config.staleDataTrendingMinutes);
+        if (!this.completedAt) return true;  // If we haven't completed it yet, do it now
+        const tootAgeInMinutes = ageInMinutes(this.tootedAt());
+        const minutesSinceCompleted = ageInMinutes(this.completedAt);
+
+        // If we have completed it, check if we need to re-evaluate for newer trending tags, links, etc.
+        return (
+            ageInMinutes(this.completedAt) > Config.staleDataTrendingMinutes  // Check if toot was completed long enough ago
+            &&
+            ageInMinutes(this.tootedAt()) < Config.tootsCompleteAfterMinutes  // But not tooted so long ago that there's little chance of new data
+        );
     }
 
     ///////////////////////////////